4 waves and star background

+++ Grame/PlayerComponent.h
@@ -3,7 +3,7 @@
 
 class PlayerComponent : public Spite::Component
 {
-	int bulletCooldown;
+	float bulletCooldown;
 public:
 	PlayerComponent(Spite::Entity* parent, Spite::ComponentID id);
 	void Update(float dt) override;
