+++ Grengine/FunctionComponent.h
@@ -3,10 +3,10 @@
 
 namespace Spite {
     class FunctionComponent : public Component {
-		std::function<void(FunctionComponent*, double)> f;
+		std::function<void(FunctionComponent*, float)> f;
 	public:
 		FunctionComponent(Entity* parent, ComponentID id);
 		void Update(float dt) override;
-		void SetFunction(const std::function<void(FunctionComponent*,double)>& function);
+		void SetFunction(const std::function<void(FunctionComponent*,float)>& function);
     };
 }
